+++ scripts/run-app
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-exec gunicorn -b "0.0.0.0:$PORT" "$@" "broker:create_app()"
+exec gunicorn -b "0.0.0.0:$PORT" "$@" "broker.app:create_app()"
