+++ tests/lib/factories.py
@@ -2,7 +2,7 @@ from factory import Sequence, SubFactory
 from factory.alchemy import SQLAlchemyModelFactory
 from openbrokerapi.service_broker import OperationState
 
-from broker import db
+from broker.extensions import db
 from broker.models import Operation, ServiceInstance, ACMEUser
